Cast the layer index where the paint clip is asked for
MSVC treats the size_t to int narrowing as an error and GCC does not, so
the Windows job caught what the local build could not. The tools hold
their layer index as size_t; App::paint_clip takes an int because the
other caller passes active_layer(), which can be -1.

+++ app/src/tools/Tools.cpp
@@ -231,7 +231,7 @@ public:
             default: break;
         }
         layer_ = app.active_layer();
-        stroke_ = std::make_unique<raster::Stroke>(app.paint_pixels(layer_), brush, color, mode, app.paint_clip(layer_));
+        stroke_ = std::make_unique<raster::Stroke>(app.paint_pixels(layer_), brush, color, mode, app.paint_clip(static_cast<int>(layer_)));
         if (mode == raster::StrokeMode::Clone || mode == raster::StrokeMode::Heal) stroke_->set_clone_source(&clone_src_, off_x_, off_y_);
         if (filter) stroke_->set_filter(std::move(filter));
         if (area_filter) stroke_->set_area_filter(std::move(area_filter));
